feat: support dynamic pricing via request headers (#32)
Enable dynamic resource pricing via a callback that
allows downstream clients to apply dynamic pricing
based on request headers. This enables flows where
reading the request body is not enough to determine
how much to charge for the request.

Note that it is up to servers to properly validate
request headers to avoid malicious requests where
users are trying to pay for less than their request
is supposed to cost.